# docs/observability/1.5/06-Java-Python内部观测接口草案.openapi.yaml
openapi: 3.0.3
info:
  title: Java-Python 内部观测 API
  version: 0.6.0-draft
  description: |
    当前环境 Java 查询当前环境 Python Task、Run、模型投影、模型生命周期和安全事件。
    测试与正式分别部署，不接受 environment 查询参数。入口必须在网关、Service Mesh
    或应用 TLS 层实际启用 mTLS；x-mtls-required 只是契约标记，代码生成不会自动启用 TLS。
servers:
  - url: /internal/observability/v1
    description: 当前环境 Python Framework 内部地址
security:
  - serviceBearer: []
...
paths:
...
  /model-invocations:
    get:
      tags: [Models]
      summary: 查询一行一个模型调度 invocation 的投影
      operationId: internalListModelInvocations
      x-mtls-required: true
      x-fixed-sort:
        - {field: startedAt, direction: DESC}
        - {field: invocationId, direction: DESC}
      description: invocation 表示进入模型调度边界的一次尝试；已确认实际请求只统计 dispatchStatus=DISPATCHED，DISPATCH_UNKNOWN 单列。首次查询物化有界结果清单，后续页不得重查可变投影。
      parameters:
        - $ref: '#/components/parameters/ScopeToken'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/DeadlineMs'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/SnapshotTo'
        - $ref: '#/components/parameters/HighWatermark'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - {name: taskId, in: query, schema: {type: string}}
        - {name: runId, in: query, schema: {type: string}}
        - {name: stageId, in: query, schema: {type: string}}
        - {name: featureCode, in: query, schema: {type: string}}
        - {name: logicalCallId, in: query, schema: {type: string}}
        - {name: invocationId, in: query, schema: {type: string}}
        - {name: provider, in: query, schema: {type: string}}
        - {name: modelPackId, in: query, schema: {type: string}}
        - {name: modelName, in: query, schema: {type: string}}
        - {name: privacyMode, in: query, schema: {type: string, enum: [STANDARD, PRIVATE]}}
        - {name: routeType, in: query, schema: {type: string, enum: [LOCAL, EXTERNAL]}}
        - {name: lifecycleStatus, in: query, schema: {$ref: '#/components/schemas/InvocationLifecycleStatus'}}
        - {name: dispatchStatus, in: query, schema: {$ref: '#/components/schemas/DispatchStatus'}}
        - {name: outcome, in: query, schema: {$ref: '#/components/schemas/Outcome'}}
        - {name: errorCode, in: query, schema: {type: string}}
      responses:
        '200': {$ref: '#/components/responses/ModelInvocationList'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /model-invocation-events:
    get:
      tags: [Models]
      summary: 查询模型追加式生命周期事件，支撑统一事件中心
      operationId: internalListModelInvocationEvents
      x-mtls-required: true
      x-fixed-sort:
        - {field: occurredAt, direction: DESC}
        - {field: eventId, direction: DESC}
      parameters:
        - $ref: '#/components/parameters/ScopeToken'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/DeadlineMs'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/SnapshotTo'
        - $ref: '#/components/parameters/HighWatermark'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - {name: taskId, in: query, schema: {type: string}}
        - {name: runId, in: query, schema: {type: string}}
        - {name: logicalCallId, in: query, schema: {type: string}}
        - {name: invocationId, in: query, schema: {type: string}}
        - {name: eventType, in: query, schema: {type: string}}
      responses:
        '200': {$ref: '#/components/responses/ModelEventList'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /model-invocations/{invocationId}:
    get:
      tags: [Models]
      summary: 获取 invocation 投影和完整生命周期事件
      operationId: internalGetModelInvocation
      x-mtls-required: true
      x-detail-consistency: LATEST
      description: 返回最新 invocation 投影及 projectionVersion/dataAsOf，并附追加式生命周期；不提供列表快照的 as-of 详情。
      parameters:
        - $ref: '#/components/parameters/ScopeToken'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/DeadlineMs'
        - {name: invocationId, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: 投影摘要和追加式生命周期
          headers:
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InternalModelInvocationDetail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFoundOrInvisible'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /model-summary:
    get:
      tags: [Models]
      summary: 在 Python/数据库投影侧聚合模型统计
      operationId: internalGetModelSummary
      x-mtls-required: true
      description: 禁止 Java 通过全量分页事件自行计算。
      parameters:
        - $ref: '#/components/parameters/ScopeToken'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/DeadlineMs'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - {name: featureCode, in: query, schema: {type: string}}
        - {name: provider, in: query, schema: {type: string}}
        - {name: modelPackId, in: query, schema: {type: string}}
        - {name: modelName, in: query, schema: {type: string}}
        - {name: privacyMode, in: query, schema: {type: string, enum: [STANDARD, PRIVATE]}}
        - {name: routeType, in: query, schema: {type: string, enum: [LOCAL, EXTERNAL]}}
      responses:
        '200':
          description: 已聚合统计
          headers:
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InternalModelSummary'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
  securitySchemes:
    serviceBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Service JWT 短期可复用并绑定 mTLS 服务身份，校验 iss、aud、exp、nbf。
        jti 用于追踪和撤销，不按一次使用处理。
  headers:
    RequestId: {description: Java 传入或 Python 生成的 request ID, schema: {type: string}}
  parameters:
    ScopeToken:
      name: X-Observability-Scope
      in: header
      required: true
      description: |
        Java 为每个下游调用单独签发的短期 JWT，绑定 method、path、queryHash、requestId、
        environment、actor、tenant scope、permissions 和 access context。每个调用使用独立 jti；
        网络重试重新签发并沿用 requestId。流式 Token 额外绑定 runId 和最大连接时长。
        这是能力型凭据，Python、网关、APM 和日志均不得记录原值。
      schema: {type: string, minLength: 20, maxLength: 4096}
    RequestId: {name: X-Request-ID, in: header, required: true, schema: {type: string}}
    Traceparent: {name: traceparent, in: header, schema: {type: string}}
    DeadlineMs:
      name: X-Request-Deadline-Ms
      in: header
      required: true
      schema: {type: integer, minimum: 100, maximum: 5000}
    From: {name: from, in: query, required: true, schema: {type: string, format: date-time}}
    To: {name: to, in: query, required: true, schema: {type: string, format: date-time}}
    SnapshotTo: {name: snapshotTo, in: query, required: true, schema: {type: string, format: date-time}}
    HighWatermark:
      name: highWatermark
      in: query
      description: 首次省略；后续页只传 Python 签发的短期、不可猜测服务端句柄（例如 hwm_01J...）。数据库 PIT、签名 JWT、租户和筛选明文只保存在 Python 服务端，不得放入 Query String
      schema: {type: string, minLength: 14, maxLength: 84, pattern: '^hwm_[A-Za-z0-9_-]{10,80}$', example: hwm_01JABCDEF1234567890}
    Cursor:
      name: cursor
      in: query
      description: 绑定 Scope、筛选、snapshotTo 和短 highWatermark 句柄的能力型游标；访问日志和 APM 必须屏蔽
      schema: {type: string, minLength: 16, maxLength: 4096}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 100}}
  responses:
...
    ModelInvocationList:
      description: 一行一个 invocation 的投影列表
      headers: {X-Request-ID: {$ref: '#/components/headers/RequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/InternalModelInvocationList'}}}
    ModelEventList:
      description: 模型生命周期事件列表
      headers: {X-Request-ID: {$ref: '#/components/headers/RequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/InternalModelEventList'}}}
...
    BadRequest: {$ref: '#/components/responses/Error400'}
    Unauthorized: {$ref: '#/components/responses/Error401'}
    Forbidden: {$ref: '#/components/responses/Error403'}
    NotFoundOrInvisible: {$ref: '#/components/responses/Error404'}
...
    TooManyRequests: {$ref: '#/components/responses/Error429'}
    InternalError: {$ref: '#/components/responses/Error500'}
    Unavailable: {$ref: '#/components/responses/Error503'}
    Error400: {description: 参数、时间窗、快照、高水位或游标无效, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Error401: {description: mTLS、Service JWT 或 Scope Token 无效, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Error403: {description: 环境、权限、Path 绑定或租户范围不允许, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Error404: {description: 资源不存在或对 Scope 不可见, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
    Error429: {description: 查询预算、速率或快照容量限制；快照容量超限使用 OBSERVABILITY_SNAPSHOT_CAPACITY_EXCEEDED, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Error500: {description: Python 内部错误, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Error503: {description: 来源不可用；Java 严格分页不得发放下一页游标, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  schemas:
    Outcome:
      type: string
      enum: [SUCCESS, FAILURE, DENIED, CANCELLED, TIMEOUT, PARTIAL, UNKNOWN, ABANDONED]
    InvocationLifecycleStatus:
      type: string
      description: RUNNING 表示仍在执行；TERMINAL 才允许填写 outcome
      enum: [RUNNING, TERMINAL]
    DispatchStatus:
      type: string
      description: 已确认实际请求只统计 DISPATCHED；DISPATCH_UNKNOWN 必须独立展示
      enum: [NOT_DISPATCHED, DISPATCHED, DISPATCH_UNKNOWN]
...
    InternalModelInvocationDetail:
      type: object
      additionalProperties: false
      required: [invocation, events]
      properties:
        invocation: {$ref: '#/components/schemas/InternalModelInvocation'}
        events: {type: array, items: {$ref: '#/components/schemas/InternalModelInvocationEvent'}}
    Percentiles:
      type: object
      additionalProperties: false
      required: [p50, p95, p99]
      properties:
        p50: {type: number, format: double, minimum: 0}
        p95: {type: number, format: double, minimum: 0}
        p99: {type: number, format: double, minimum: 0}
    BreakdownItem:
      type: object
      additionalProperties: false
      required: [key, count, ratio]
      properties:
        key: {type: string}
        count: {type: integer, format: int64, minimum: 0}
        ratio: {type: number, format: double, minimum: 0, maximum: 1}
    InternalModelSummary:
      type: object
      additionalProperties: false
      required: [logicalCallCount, dispatchAttemptCount, dispatchedRequestCount, notDispatchedCount, dispatchUnknownCount, activeAttemptCount, logicalSuccessRate, attemptSuccessRate, retryRate, costs]
      properties:
        logicalCallCount: {type: integer, format: int64, minimum: 0}
        dispatchAttemptCount: {type: integer, format: int64, minimum: 0}
        dispatchedRequestCount: {type: integer, format: int64, minimum: 0}
        notDispatchedCount: {type: integer, format: int64, minimum: 0}
        dispatchUnknownCount: {type: integer, format: int64, minimum: 0}
        activeAttemptCount: {type: integer, format: int64, minimum: 0, description: lifecycleStatus=RUNNING 的调度尝试，不进入终态成功率分母}
        logicalSuccessRate: {type: number, format: double, minimum: 0, maximum: 1}
        attemptSuccessRate: {type: number, format: double, minimum: 0, maximum: 1, description: 分母只包含已终态且 DISPATCHED 的 invocation}
        retryRate: {type: number, format: double, minimum: 0, maximum: 1, description: NOT_DISPATCHED 不计模型重试}
        inputTokens: {type: integer, format: int64, minimum: 0}
        outputTokens: {type: integer, format: int64, minimum: 0}
        costs: {type: array, items: {$ref: '#/components/schemas/CurrencyAmount'}}
        latencyPercentiles: {$ref: '#/components/schemas/Percentiles'}
        ttftPercentiles: {$ref: '#/components/schemas/Percentiles'}
        timeSeries: {type: array, items: {$ref: '#/components/schemas/ModelTimeSeriesPoint'}}
        errorBreakdown: {type: array, items: {$ref: '#/components/schemas/BreakdownItem'}}
        providerBreakdown: {type: array, items: {$ref: '#/components/schemas/BreakdownItem'}}
        modelBreakdown: {type: array, items: {$ref: '#/components/schemas/BreakdownItem'}}
        routeTypeBreakdown: {type: array, items: {$ref: '#/components/schemas/BreakdownItem'}}
        privacyModeBreakdown: {type: array, items: {$ref: '#/components/schemas/BreakdownItem'}}
    ModelTimeSeriesPoint:
      type: object
      additionalProperties: false
      required: [bucketStart, logicalCallCount, dispatchAttemptCount, dispatchedRequestCount, successRate, p95LatencyMs]
      properties:
        bucketStart: {type: string, format: date-time}
        logicalCallCount: {type: integer, format: int64, minimum: 0}
        dispatchAttemptCount: {type: integer, format: int64, minimum: 0}
        dispatchedRequestCount: {type: integer, format: int64, minimum: 0}
        successRate: {type: number, format: double, minimum: 0, maximum: 1}
        p95LatencyMs: {type: number, format: double, minimum: 0}
    CurrencyAmount:
      type: object
      additionalProperties: false
      required: [currency, amount]
      properties:
        currency: {type: string, pattern: '^[A-Z]{3}$'}
        amount: {type: number, format: double, minimum: 0}
    PageInfo:
      type: object
      additionalProperties: false
      required: [hasMore, limit]
      properties:
        nextCursor: {type: string, nullable: true}
        hasMore: {type: boolean}
        limit: {type: integer, minimum: 1}
...
    InternalModelInvocationList:
      type: object
      additionalProperties: false
      required: [data, page, watermark]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/InternalModelInvocation'}}
        page: {$ref: '#/components/schemas/PageInfo'}
        watermark: {$ref: '#/components/schemas/SourceWatermark'}
    InternalModelEventList:
      type: object
      additionalProperties: false
      required: [data, page, watermark]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/InternalModelInvocationEvent'}}
        page: {$ref: '#/components/schemas/PageInfo'}
        watermark: {$ref: '#/components/schemas/SourceWatermark'}
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, requestId]
      properties:
        code: {type: string}
        requestId: {type: string}
        retryable: {type: boolean}